CLI: refactor, move CLI related stuff to its own module
